% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{acc_fun}
\alias{acc_fun}
\title{Accuracy of model classification}
\usage{
acc_fun(x, y, ntot)
}
\arguments{
\item{x}{Numeric value for the isotope value of the synthetic value.}

\item{y}{p value.}

\item{ntot}{Numeric for total number of model runs.}
}
\value{
Numeric vector for accuracy.
}
\description{
This function can be used in \code{dplyr::\link[dplyr:summarise]{summarise}}
to calculate the accuracy of model classification for the intra- and inter-
isotope test.
}
