% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{demo}
\format{
A data frame with 1 row and 1 variable:
\describe{
  \item{message}{message}
}
}
\usage{
demo
}
\description{
A dataset containing a message describing how to obtain the data for this
study
}
\keyword{datasets}
