% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_64_MON.R
\docType{data}
\name{diag_64_MON}
\alias{diag_64_MON}
\title{diag_64_MON}
\format{
A data frame with 3 rows and 5 variables:
\describe{
  \item{dim.nm}{dimension name}
  \item{`$\\hat{\\bar{R}}$`}{latex formatted grand mean R}
  \item{`$\hat{\epsilon}_{\bar{R}}$ (\text{\textperthousand})`}{Relative standard deviation}
  \item{`$\\Delta AIC_{\bar{R}}`}{Difference in AIC}
  \item{`$p_{\\bar{R}}$`}{p value}
}
}
\usage{
diag_64_MON
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
