% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_point.R
\name{download_point}
\alias{download_point}
\alias{write_point}
\alias{save_point}
\title{Write point ion count data}
\usage{
download_point(type = "processed", on_build = FALSE)

write_point(obj, on_build = FALSE)

save_point(ggplot, name, width, height, unit, on_build = FALSE)
}
\description{
A wrapper around zen4R in order to download the large matlab files as well as processed data cubes.
}
