% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matlab.R
\name{get_matlab}
\alias{get_matlab}
\title{Get path to matlab file}
\usage{
get_matlab(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
This function comes from the package `readr`, and has been modified to access
the bundled datasets in directory `inst/extdata` of `pointapply`. This
function make them easy to access. This function is modified from
\code{\link[readr:readr_example]{readr_example}} of the package
\code{\link[readr]{readr}}.
}
\examples{
get_matlab()
get_matlab("2020-08-20-GLENDON")
}
