% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_IC.R
\name{gg_cnts}
\alias{gg_cnts}
\title{Raster images for ion count ratios}
\usage{
gg_cnts(
  title,
  ion1,
  ion2,
  viri = "D",
  res = 256,
  grid_cell = 64,
  scaler = 40/256,
  compilation = FALSE,
  save = FALSE,
  .name = "map_raster_image"
)
}
\arguments{
\item{title}{Character string for the analyte ("MEX" or "MON") to be used.}

\item{ion1}{A character string constituting the rare isotope ("13C").}

\item{ion2}{A character string constituting the common isotope ("12C").}

\item{viri}{Character string selecting the viriditas color scheme.}

\item{res}{Pixel resolution.}

\item{grid_cell}{Pixel size of grid_cells (one side of the square).}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}

\item{compilation}{Logical whether to plot all three dimensions according to
the grid layout
(see \code{pointapply::\link[pointapply:gg_sketch]{gg_sketch}}).}

\item{save}{Boolean whether to save the plot as an png.}

\item{.name}{Name of the raster file.}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_cnts} Raster images based on the ratio of ion counts.
}
