% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_IC.R
\name{gg_effect}
\alias{gg_effect}
\alias{gg_sketch}
\title{Combination plot}
\usage{
gg_effect(
  ls,
  ls_im,
  ttl,
  ratio,
  grid_print = FALSE,
  viri = "A",
  res = 256,
  grid_cell = 64,
  scaler = 40/256,
  Xt = "Xt.pr",
  grid = "grid.nm"
)

gg_sketch(res = 256, grid_cell = 64, scaler = 40/256)
}
\arguments{
\item{ls}{Named list containing dataframes for planes of ion counts for
high precision isotope ratios.}

\item{ls_im}{Named list containing dataframes for planes of ion counts for
ion raster map.}

\item{ttl}{Character string or expression for plot title.}

\item{ratio}{A character string constituting the ion ratio, where the two
ions are separated by a dash ("12C14N-40Ca16O").}

\item{grid_print}{Logical whether to print the grid numbers
(default = FALSE).}

\item{viri}{Character string selecting the viriditas color scheme
(default = "A").}

\item{res}{Resolution of ion map in pixels (default = 256).}

\item{grid_cell}{Grid-cell size in pixels (default = 64).}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}

\item{Xt}{Variable for ion count rates (default = Xt.pr).}

\item{grid}{Variable for grid_cell number (default = grid.nm).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_effect} combines raster images for ion count ratios with high
precision isotope ratios. \code{gg_sketch} draws the layout of the
grid_cells.
}
