% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indepth_IC.R
\name{gg_point}
\alias{gg_point}
\title{In-depth analysis of IC data cubes}
\usage{
gg_point(
  title,
  grid_cell,
  ion1_thr,
  ion2_thr,
  thr,
  ion1_R,
  ion2_R,
  colors = c("#8E063B", "#023FA5"),
  save = FALSE,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL
)
}
\arguments{
\item{title}{Character string for the analyte ("MEX" or "MON") to be used.}

\item{ion1_thr}{Character string for ion in the enumerator of ion ratio of
the map.}

\item{ion2_thr}{Character string for ion in the denominator of ion ratio of
the map.}

\item{thr}{Numeric threshold value for filter selection.}

\item{ion1_R}{A character string constituting the rare isotope ("13C").}

\item{ion2_R}{A character string constituting the common isotope ("12C").}

\item{colors}{Colour palette to identify matrix and inclusion (default =
\code{ c("#8E063B", "#023FA5")}).}

\item{save}{Boolean whether to save the plot as an png.}

\item{.X}{Variable for ion count rates (default = NULL).}

\item{.N}{Variable for ion counts (default = NULL).}

\item{.species}{Variable for species names (default = NULL).}

\item{.t}{Variable for time increment (default = NULL).}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_point} This plot function is used to produce the scatter plots of
Figure 8 and Supplementary Figure 9.
}
