% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heat_map}
\alias{heat_map}
\alias{heat_sum}
\title{Heatmaps and summary plots for model performance.}
\usage{
heat_map(
  simu,
  x,
  y,
  stat,
  grp1,
  grp2,
  conversion,
  ttl,
  x_lab,
  y_lab,
  x_sec = NULL,
  trans_base = -2,
  trans_n = 10
)

heat_sum(
  simu,
  x,
  y,
  stat,
  grp1,
  grp2,
  conversion,
  ttl,
  x_lab,
  y_lab,
  x_sec,
  y_sec,
  trans_base = -20,
  trans_n = 6
)
}
\arguments{
\item{simu}{Ion count data.}

\item{x}{Variable (continuous variable number one) isotope offset in paper.}

\item{y}{Variable (continuous variable number two) excess ionization
efficiency in paper.}

\item{stat}{Numeric for the model accuracy (between 0 and 1).}

\item{grp1}{Nominal factorial number one; outlier detection method in paper.}

\item{grp2}{Nominal factorial number one; for intra-isotope variability in
paper.}

\item{conversion}{Named vector for the conversion of the ionization trend to
excess ionization efficiency.}

\item{ttl}{Character string or expression for the plot title.}

\item{x_lab}{Character string or expression for the x-axis label.}

\item{y_lab}{Character string or expression for the y-axis label.}

\item{x_sec}{Character string or expression for the secondary x-axis label.}

\item{trans_base}{Isotope value for the scale transformation for fractional
size of the isotope anomaly.}

\item{trans_n}{Initial sample size for the scale transformation for
fractional size of the isotope anomaly.}

\item{y_sec}{Character string or expression for the secondary y-axis label.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{heat_map} generates a heatmap for model accuracy evaluation along
two continuous variables and up to two categorical values. Instead
\code{heat_sum} summarise the heatmap over opne of the two continuous and
crossed variables.
}
