% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_class.R
\name{hyp_class}
\alias{hyp_class}
\title{Barplots for hypothesis testing}
\usage{
hyp_class(IC, sec_vc, ttl, save = FALSE)
}
\arguments{
\item{IC}{Ion count data}

\item{sec_vc}{Named vector for secondary axis.}

\item{ttl}{Character string or expression for plot title.}

\item{save}{Boolean whether to save the plot as an png.}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{hyp_class} Barplot for the classification of failing to reject a zero
hypothesis (Fig. 3 and Supplementary Fig. 5)
}
