% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sum_grid_64_all.R
\docType{data}
\name{map_sum_grid_64_all}
\alias{map_sum_grid_64_all}
\title{map_sum_grid_64_all}
\format{
A data frame with 33,760 rows and 10 variables:
\describe{
  \item{dim.nm}{dimension name}
  \item{grid.nm}{grid-cell}
  ...
}
}
\usage{
map_sum_grid_64_all
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
