% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointdata.R
\docType{data}
\name{pointdata}
\alias{pointdata}
\title{pointdata}
\format{
An object of class \code{character} of length 1.
}
\usage{
pointdata
}
\description{
Data is stored externally on Zenodo, use pointapply::download_point()
}
\keyword{datasets}
