% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matlab.R
\name{read_matlab}
\alias{read_matlab}
\title{Read matlab ion count data}
\usage{
read_matlab(
  directory,
  plane,
  title,
  species = NULL,
  grid_cell = 64,
  output = "sum",
  grid_sel = NULL,
  scaler = 40/256
)
}
\arguments{
\item{directory}{Character string for path to the Matlab data-file}

\item{plane}{Variable for the dimension over which to be aggregated}

\item{grid_cell}{Integer indicating size of cells in pixels, only exponent of
base two allowed}

\item{output}{Character string for type of output; \code{"complete"} or
\code{"sum"}.}

\item{grid_sel}{Integer selecting an grid-cell for in-depth analysis
(default is 64 pixels as used in this study).}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing the ion counts
aggregated over the plane of choice, and if \code{output = "sum"}.
}
\description{
\code{read_matlab} Tabulate  and aggregate the ion count data contained in
matlab files exported by LANS.
}
\examples{
# Use get_matlab() to path to data bundled with this package

# Pre-process real data as in the paper; Schobben, Kienhuis, Polerecky 2021
tb_rw <- read_matlab(get_matlab("2020-08-20-GLENDON"), plane = height,
                     title = "MEX")
}
