% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_paper.R
\name{render_paper}
\alias{render_paper}
\alias{edit_paper}
\title{Render the paper}
\usage{
render_paper(
  title = "Schobbenetal_SIMS_method",
  output_dir = fs::path_wd(),
  type_ms = "preprint",
  copy_figures = TRUE
)

edit_paper(type_ms = "preprint")
}
\arguments{
\item{title}{Character string for title of rendered document.}

\item{output_dir}{Character string for directory containing the rendered
document.}

\item{type_ms}{Character string for type manuscript, either \code{"preprint"}
or \code{"paper"}.}

\item{copy_figures}{Logical wether to copy figures to directory
(default = TRUE).}
}
\value{
Rendered documents including a tex, docx and pdf file. Depending on
the argument \code{copy_figures} the call returns figures in png format
within its own directory (figures).
}
\description{
\code{render_paper()} is a wrapper around
\code{rmarkdown::\link[rmarkdown:render_site]{render_site}} which renders the
paper of this study.
}
