% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_coder.R
\name{sig_coder}
\alias{sig_coder}
\title{Create significance codes for p values}
\usage{
sig_coder(x = NULL, make_lab = TRUE)
}
\arguments{
\item{x}{p value (numeric) (default = NULL)}

\item{make_lab}{Logical whether to return a legend for the significance
stars (default = TRUE).}
}
\value{
Character string encoding for the significance star associated with
the p value. If argument \code{make_lab = TRUE} then a list is return, where
the second element contains the legend.
}
\description{
The function \code{sig_coder()} produces significance stars associated with p
values
}
