% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_X_64_all.R
\docType{data}
\name{stat_X_64_all}
\alias{stat_X_64_all}
\title{stat_X_64_all}
\format{
A data frame with 96 rows and 15 variables:
\describe{
  \item{file.nm}{file name}
  \item{sample.nm}{sample name}
  ...
}
}
\usage{
stat_X_64_all
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
