% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_QSA_64_all.R
\docType{data}
\name{test_QSA_64_all}
\alias{test_QSA_64_all}
\title{test_QSA_64_all}
\format{
A data frame with 12,784 rows and 26 variables:
\describe{
  \item{file.nm}{file name}
  \item{sample.nm}{sample name}
  ...
}
}
\usage{
test_QSA_64_all
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
