% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICtheme.R
\name{themes_IC}
\alias{themes_IC}
\title{Theming for pointapply plots}
\usage{
themes_IC(base = ggplot2::theme_classic(), ...)
}
\arguments{
\item{base}{Default theme.}

\item{...}{Currently not supported.}
}
\value{
Ggplot2 theme.
}
\description{
Theming for pointapply plots
}
