% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twodens.R
\name{twodens}
\alias{twodens}
\title{Scatter plots for very dens datasets}
\usage{
twodens(
  .data,
  x,
  y,
  xlab,
  ylab,
  ttl,
  downsample,
  gr,
  x_lim = NULL,
  y_lim = NULL,
  unit = "um",
  sds = 22,
  geom = "point",
  facet_sc = "fixed"
)
}
\arguments{
\item{.data}{Dataframe.}

\item{x}{Variable for the x axis.}

\item{y}{Variable for the y axis.}

\item{xlab}{Character string or expression for x axis label.}

\item{ylab}{Character string or expression for x axis label.}

\item{ttl}{Character string or expression for plot title.}

\item{downsample}{A numeric supplying the fraction for downsampling of the original
dataframe.}

\item{gr}{A grouping variable.}

\item{x_lim}{Numeric range for x axis (default = NULL).}

\item{y_lim}{Numeric range for y axis (default = NULL).}

\item{unit}{A character string for the units of the grouping variable. The
string \code{"um"} is used for the grid cell size comparison plots. The
string \code{"dim"} is used for all other plots, and defaults to the grouping
variable name.}

\item{sds}{Numeric settings the seeds for downsampling via `set.seed()`.}

\item{geom}{A character string for the geometry mapped on the aesthetics.
Options are  \code{"point"} for normal scatter plots, and \code{"dens2d"} for
density contour lines.}

\item{facet_sc}{The \code{scales} parameter of `ggplot2::facet_grid`}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{twodens} Scatter plots with 2D density as color and alpha scale to
visualise very dens data
}
