% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_64_MON.R
\docType{data}
\name{diag_64_MON}
\alias{diag_64_MON}
\title{diag_64_MON}
\format{
A data frame with 33,760 rows and 10 variables:
\describe{
  \item{dim.nm}{dimension name}
  \item{`$\\hat{\\bar{R}}$`}{latex formatted grand mean R}
  ...
}
}
\usage{
diag_64_MON
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
