% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_IC.R
\name{gg_cnts}
\alias{gg_cnts}
\title{Raster images for ion count ratios}
\usage{
gg_cnts(
  IC,
  ion1,
  ion2,
  ttl,
  viri,
  height = 256,
  width = 256,
  grid_cell = 64,
  scaler = 40/256,
  compilation = FALSE
)
}
\arguments{
\item{IC}{Ion count dataframe.}

\item{ion1}{A character string constituting the rare isotope ("13C").}

\item{ion2}{A character string constituting the common isotope ("12C").}

\item{ttl}{Character string or expression for plot title.}

\item{viri}{Character string selecting the viriditas color scheme.}

\item{height}{Pixels of height.}

\item{width}{Pixels of width.}

\item{grid_cell}{Pixel size of grid_cells (one side of the square).}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}

\item{compilation}{Logical whether to plot all three dimensions according to
the grid layout
(see \code{pointapply::\link[pointapply:gg_sketch]{gg_sketch}}).}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_cnts} Raster images based on the ratio of ion counts.
}
