% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_IC.R
\name{gg_effect}
\alias{gg_effect}
\alias{gg_sketch}
\title{Combination plot}
\usage{
gg_effect(
  title,
  ratio,
  grid_print = FALSE,
  viri = "A",
  res = 256,
  grid_cell = 64,
  scaler = 40/256,
  label = "latex",
  save = FALSE,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .ion1 = "13C",
  .ion2 = "12C"
)

gg_sketch(grid_cell = 64, res = 256, scaler = 40/256)
}
\arguments{
\item{title}{Character string for the analyte ("MEX" or "MON") to be used.}

\item{ratio}{A character string constituting the ion ratio, where the two
ions are separated by a dash ("12C14N-40Ca16O").}

\item{grid_print}{Logical whether to print the grid numbers
(default = FALSE).}

\item{viri}{Character string selecting the viriditas colour scheme
(default = "A").}

\item{res}{Resolution of ion map in pixels (default = 256).}

\item{grid_cell}{Grid-cell size in pixels (default = 64).}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}

\item{label}{Character string indicating how to name tibble columns
(default = "latex").}

\item{save}{Boolean whether to save the plot as an png.}

\item{.X}{Variable for ion count rates (default = Xt.pr).}

\item{.N}{Variable for ion counts (default = N.pr).}

\item{.species}{Variable for species names (default = species.nm).}

\item{.t}{Variable for time increment (default = t.nm).}

\item{.ion1}{Character string for rare isotope (default = "13C").}

\item{.ion2}{Character string for common isotope (default = "12C").}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_effect} combines raster images for ion count ratios with high
 precision isotope ratios. \code{gg_sketch} draws the layout of the
 grid_cells.
}
