% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indepth_IC.R
\name{gg_point}
\alias{gg_point}
\title{In-depth analysis of one grid_cell}
\usage{
gg_point(
  IC,
  image,
  ion1_thr,
  ion2_thr,
  thr,
  ion1_R,
  ion2_R,
  Xt = Xt.pr,
  N = N.rw,
  colors = c("#FFEDA0", "#FEB24C")
)
}
\arguments{
\item{IC}{Ion count data.}

\item{image}{Data frame for ion raster map.}

\item{ion1_thr}{Character string for ion in the enumerator of ion ratio of
the map.}

\item{ion2_thr}{Character string for ion in the denominator of ion ratio of
the map.}

\item{thr}{Numeric threshold value for filter selection.}

\item{ion1_R}{A character string constituting the rare isotope ("13C").}

\item{ion2_R}{A character string constituting the common isotope ("12C").}

\item{Xt}{Variable for ion count rates (default = Xt.pr).}

\item{N}{Variable for ion counts (default = N.rw).}

\item{colors}{Color palette to identify matrix and inclusion (default =
\code{ c("#FFEDA0", "#FEB24C")}).}
}
\value{
\code{ggplot2::\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{gg_effect} This plot function is used to produce the scatterplots of
Figure 9 and Supplementary Figure 8.
}
