% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_class.R
\name{hyp_class}
\alias{hyp_class}
\title{Barplots for hypothesis testing}
\usage{
hyp_class(.data, sec_vc, ttl)
}
\arguments{
\item{.data}{Dataframe.}

\item{sec_vc}{Vector for secondary axis.}

\item{ttl}{Character string or expression for plot title.}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot}}.
}
\description{
\code{hyp_class} Barplot for the classification of failing to reject a zero
hypothesis (Fig. 3 and Supplementary Fig. 5)
}
