% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matlab.R
\name{read_matlab}
\alias{read_matlab}
\title{Read matlab ion count data}
\usage{
read_matlab(
  directory,
  plane,
  title,
  species = NULL,
  grid_cell = 64,
  output = "sum",
  grid_sel = NULL,
  scaler = 40/256
)
}
\arguments{
\item{directory}{Character string for path to the Matlab data-file}

\item{plane}{Variable for the dimension over which to be aggregated}

\item{title}{Character string for sample name.}

\item{species}{Character string or vector with chemical species.}

\item{grid_cell}{Integer indicating size of cells in pixels, only exponent of
base two allowed (default = 64)}

\item{output}{Character string for type of output; \code{"complete"} or
\code{"sum"}.}

\item{grid_sel}{Integer selecting an grid-cell for in-depth analysis.}

\item{scaler}{Numeric converting the pixels to metric dimension of
measurement (default is the conversion used in this study).}
}
\value{
A \code{tibble::\link[tibble:tibble]{tibble}} containing the ion counts
aggregated over the plane of choice, if \code{output = "sum"}. If
\code{output = "complete"} only works in combination with \code{grid_sel} and
produces a pixel-by-pixel dataset for that grid-cell.
}
\description{
\code{read_matlab} Tabulate  and aggregate the ion count data contained in
matlab files exported by LANS. Matlab files are read with
\code{R.matlab::\link[R.matlab:readMat]{readMat}}.
}
